Re-enable link to HACKING.rst.
It was removed because I though makes pypi failing to format the
README.rst file. It turned out that it is completely something else
(pypi formatter adding '>' in front of the 'From source' section
title...). Since I can edit the readme to fix this kind of error
directly on pypi website, it is better to re-enable this link so that
github visitor can use it.